refactor(frontend): frontend refactored (#29)
* fix(frontend): events on mobile on mainpage fix

* feat(frontend): participants title on mainpage add

* refactor(frontend): frontend refactored, now code more readable and clean

* fix(frontend): faq fixed
