+++ frontend/src/components/MainPage/FAQSection.jsx
@@ -1,0 +1,114 @@
+import React from 'react';
+import {
+    Accordion,
+    AccordionSummary,
+    AccordionDetails,
+    Box,
+    Container,
+    Link,
+    Stack,
+    Typography
+} from "@mui/material";
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+
+function FAQSection() {
+    const faqItems = [
+        {
+            question: "Как начать пользоваться FindMyTeam?",
+            answer: "Перейдите на главную страницу, создайте аккаунт или войдите. После входа вы сможете выбрать мероприятие, зарегистрироваться и указать свою роль — участник или тимлид."
+        },
+        {
+            question: "Как зарегистрироваться на мероприятие?",
+            answer: "В разделе \"Мероприятия\" выберите интересующее событие и нажмите \"Я участвую\". Укажите резюме, выберите трек и роль. Если вы тимлид, создайте команду и добавьте вакансии. Вы также можете добавить или убрать вакансии позже на странице команды. В резюме, описании команд и описании вакансий поддерживается Markdown разметка."
+        },
+        {
+            question: "Как изменить резюме, трек или роль на мероприятии?",
+            answer: "Для этого перейдите на страницу мероприятия, затем нажмите на иконку \"Карандаш\" рядом с данными, которые хотите редактировать (данные об участии находятся внизу страницы). Вы можете изменить резюме и трек, а также отказаться от участия в мероприятии. Чтобы сменить роль с \"Тимлид\" на \"Участник\", выйдете из команды или удалите команду. Если тимлид выходит из команды, команда автоматически удаляется, а тимлид становится обычным участником. Чтобы сменить роль с \"Участник\" на \"Тимлид\" откажитесь от участия в мероприятии, а затем зарегистрируйтесь заново, выбрав роль \"Тимлид\" на этапе регистрации. Обратите внимание, что в таком случае все данные об участии в мероприятии (резюме и трек) удалятся."
+        },
+        {
+            question: "Как собрать команду?",
+            answer: "Тимлид создаёт команду при регистрации, добавляет вакансии и может приглашать участников через вкладку \"Участники\". Участники могут откликаться на вакансии команд с подходящим треком."
+        },
+        {
+            question: "Как посмотреть свои мероприятия?",
+            answer: "Перейдите в личный кабинет. Там отображаются все мероприятия, где вы участвуете, с вашими ролями и треками."
+        },
+        {
+            question: "Что делать, если я не вижу подходящую команду?",
+            answer: "Можно создать собственную команду, если вы тимлид, или подождать, пока откроются новые вакансии. Также вы можете откликнуться на несколько подходящих вакансий сразу."
+        },
+        {
+            question: "Как отправить приглашение?",
+            answer: "Отправить приглашение можно только если вы тимлид и в вашей команде есть открытые вакансии. Откройте список участников на странице мероприятия, нажмите на кнопку \"пригласить\" на карточке участника, а затем выберите вакансию, на которую вы хотите его пригласить. Посмотреть приглашения, которые вы отправили можно во вкладке \"Мои приглашения\" на странице мероприятия."
+        },
+        {
+            question: "Как принять приглашение в команду?",
+            answer: "Все приглашения отображаются во вкладках \"Приглашения мне\" на странице мероприятия. Вы можете принять или отклонить приглашения в один клик."
+        },
+        {
+            question: "Как посмотреть участников команды?",
+            answer: "Чтобы посмотреть подробную информацию о команде, откройте вкладку \"команды\" на странице мероприятий. Далее нажмите на строку с командой, чтобы перейти на страницу этой команды."
+        },
+        {
+            question: "Как управлять своей командой?",
+            answer: "Управление командной доступно только если вы тимлид в этой команде. Для этого перейдите на страницу с вашей командой. Там вы можете редактировать название и описание команды, вы также можете добавлять и удалять вакансии. Для удаления вакансий нажмите на иконку мусорной корзины на карточке вакансии. Для удаления участников из команды нажмите на кнопку \"выгнать\" на карточке участника."
+        },
+        {
+            question: "Как откликнуться на вакансию?",
+            answer: "Откликнуться на вакансию можно только если вы участник мероприятия и у вас еще нет команды. Чтобы откликнуться на вакансию, откройте вкладку \"вакансии\" на странице мероприятия. Далее нажмите \"откликнуться\" на карточке вакансии, которая вам интересна. Вы также можете это сделать со страницы команды."
+        },
+        {
+            question: "Как редактировать данные в профиле или выйти из аккаунта?",
+            answer: "В личном кабинете вы можете изменить имя, e-mail, пароль и Telegram. Там же можно выйти или удалить профиль. Если вы хотите изменить резюме или отказаться от участия в каком-то мероприятии, вы можете сделать это на странице мероприятия."
+        }
+    ];
+
+    return (
+        <Box sx={{ py: 7, bgcolor: "grey.900", color: "text.primary" }}>
+            <Container maxWidth="md">
+                <Typography
+                    variant="h4"
+                    align="center"
+                    gutterBottom
+                    sx={{ fontWeight: 700, mb: 4 }}
+                >
+                    Часто задаваемые вопросы (FAQ)
+                </Typography>
+
+                <Stack>
+                    {faqItems.map((item, idx) => (
+                        <Accordion key={idx}>
+                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
+                                <Typography variant="subtitle1" fontWeight={600}>
+                                    {item.question}
+                                </Typography>
+                            </AccordionSummary>
+                            <AccordionDetails>
+                                <Typography variant="body2">
+                                    {item.answer}
+                                    {item.answerLink && (
+                                        <>
+                                            {' '}
+                                            <Link
+                                                href={item.answerLink}
+                                                target="_blank"
+                                                rel="noopener noreferrer"
+                                                sx={{ textDecoration: 'none' }}
+                                            >
+                                                {item.answerLinkText}
+                                            </Link>
+                                            {item.answerSuffix}
+                                        </>
+                                    )}
+                                </Typography>
+                            </AccordionDetails>
+                        </Accordion>
+                    ))}
+                </Stack>
+            </Container>
+        </Box>
+    );
+}
+
+export default FAQSection;
+
