+++ frontend/src/pages/MainPage.jsx
@@ -1,43 +1,21 @@
 import React, { useEffect, useState } from "react";
 import {
     Box,
-    Button,
-    ButtonGroup,
-    Card,
-    CardContent,
-    Chip,
-    Container,
-    Grid,
-    Link, Skeleton,
-    Stack,
-    Toolbar,
-    Typography,
-    useMediaQuery,
-    useTheme,
-    Accordion,
-    AccordionSummary,
-    AccordionDetails,
+    Toolbar
 } from "@mui/material";
-import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos";
-import EventIcon from "@mui/icons-material/Event";
-import GroupAddIcon from "@mui/icons-material/GroupAdd";
-import EmojiEventsIcon from "@mui/icons-material/EmojiEvents";
-import HelpOutlineIcon from "@mui/icons-material/HelpOutline";
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { useNavigate } from "react-router-dom";
 import {apiFetch, refreshAccessToken} from "../apiClient.js";
-import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
-import {grey} from "@mui/material/colors";
+import HeroSection from "../components/MainPage/HeroSection.jsx";
+import EventsSection from "../components/MainPage/EventsSection.jsx";
+import SupportedEventsSection from "../components/MainPage/SupportedEventsSection.jsx";
+import FAQSection from "../components/MainPage/FAQSection.jsx";
+import Footer from "../components/MainPage/Footer.jsx";
+
 function MainPage() {
     const navigate = useNavigate();
-    const theme = useTheme();
-    const isXs = useMediaQuery(theme.breakpoints.down("sm"));
-
     const [events, setEvents] = useState([]);
     const [isLoggedIn, setIsLoggedIn] = useState(false);
     const [loading, setLoading] = useState(true);
-
-
     const [participations, setParticipations] = useState([]);
 
     useEffect(() => {
@@ -64,27 +42,11 @@ function MainPage() {
             } catch (err) {
                 console.error("Ошибка загрузки:", err);
             } finally {
-                setLoading(false); // окончание загрузки
+                setLoading(false);
             }
         }
         fetchData();
     }, []);
-
-
-
-    useEffect(() => {
-        setLoading(true);
-        try {
-        apiFetch("/events")
-            .then((res) => res.json())
-            .then((data) => setEvents(data || []))
-            .catch((err) => console.error("Ошибка загрузки событий:", err))
-
-        }
-        finally {
-            setLoading(false);
-        }
-    }, [navigate]);
 
     useEffect(() => {
         refreshAccessToken()
@@ -92,670 +54,23 @@ function MainPage() {
             .catch(() => setIsLoggedIn(false));
     }, []);
 
-    // Общий стилевой Section - занимает высоту экрана, но не больше. при избытке контента прокрутки нет
-    const sectionSx = {
-        height: "100svh",
-        maxHeight: "100svh",
-        display: "grid",
-        gridTemplateRows: "1fr",
-        alignItems: "center",
-        // overflowY: "auto",
-        py: { xs: 6, md: 8 },
-    };
-
     const participationMap = participations.reduce((acc, p) => {
         acc[p.event_id] = p;
         return acc;
     }, {});
 
-    // Кнопки вплотную/ на xs — вертикально вплотную, на sm — горизонтально вплотную
-    const AuthButtons = ({ loggedIn }) => (
-        <Box
-        >
-            {!loggedIn ? ( <ButtonGroup variant="contained" aria-label="Basic button group">
-                <Button variant="outlined" sx={{width: {sm: "120px" }}} onClick={() => navigate("/auth")}>Войти</Button>
-                    <Button variant="contained" onClick={() => navigate("/events")}>Доступные мероприятия</Button>
-                </ButtonGroup>
-            ) : (
-                <ButtonGroup variant="contained" aria-label="Authenticated user buttons">
-                    <Button variant="contained" onClick={() => navigate("/home")}>Мой профиль</Button>
-                    <Button variant="contained" onClick={() => navigate("/events")}>Доступные мероприятия</Button>
-                </ButtonGroup> )}
-
-        </Box>
-    );
-
-    const renderSkeletonCard = () => (
-        <Card sx={{ mt: 3, borderRadius: 3 }}>
-            <CardContent>
-                <Skeleton variant="text" width="60%" height={30} />
-                <Skeleton variant="text" width="80%" />
-                <Skeleton variant="rectangular" width="100%" height={80} sx={{ mt: 2, borderRadius: 1 }} />
-                <Skeleton variant="text" width="40%" sx={{ mt: 2 }} />
-                <Skeleton variant="rectangular" width={120} height={36} sx={{ mt: 2, borderRadius: 2 }} />
-            </CardContent>
-        </Card>
-    );
-
     return (
         <>
             <Toolbar />
-
-            <Box
-                sx={{
-                    ...sectionSx,
-                    position: "relative",
-                    color: "white",
-                    textAlign: "center",
-                    overflow: "hidden",
-                    height: "100vh",     // секция на всю высоту экрана
-                }}
-            >
-                {/* Видео: десктоп */}
-                <Box
-                    sx={{
-                        position: "absolute",
-                        inset: 0,           // top:0 right:0 bottom:0 left:0
-                        zIndex: 0,
-                        display: { xs: "none", md: "block" },
-                        pointerEvents: "none",
-                    }}
-                >
-                    <video
-                        muted
-                        playsInline
-                        autoPlay
-                        style={{
-                            width: "100%",
-                            height: "100%",
-                            objectFit: "cover",
-                            display: "block",
-                        }}
-                    >
-                        <source src="/video_2.mp4" type="video/mp4" />
-                    </video>
-                </Box>
-
-                {/* Видео: мобилка */}
-                <Box
-                    sx={{
-                        position: "absolute",
-                        inset: 0,
-                        zIndex: 0,
-                        display: { xs: "block", md: "none" },
-                        pointerEvents: "none",
-                    }}
-                >
-                    <video
-                        muted
-                        playsInline
-                        autoPlay
-                        loop
-                        style={{
-                            width: "100%",
-                            height: "100%",
-                            objectFit: "cover",
-                            display: "block",
-                        }}
-                    >
-                        <source src="/video_mobile.mp4" type="video/mp4" />
-                    </video>
-                </Box>
-
-                {/* Контент поверх видео */}
-                <Box
-                    sx={{
-                        position: "absolute",
-                        inset: 0,
-                        zIndex: 1,
-                        display: "grid",
-                        placeItems: "center",     // идеальное центрирование
-                        px: 2,                    // боковые поля для мобайла
-                    }}
-                >
-                    <Container maxWidth="md">
-                        <Box sx={{ textAlign: "center" }}>
-                            <Typography
-                                variant={isXs ? "h4" : "h3"}
-                                gutterBottom
-                                sx={{
-                                    color: "primary.main",
-                                    fontWeight: 800,
-                                    letterSpacing: ".02em",
-
-                                    // textShadow: "0 0 5px rgba(0,229,255,.7), 0 0 24px rgba(0,229,255,.35)",
-                                }}
-                            >
-                                Добро пожаловать в FindMyTeam
-                            </Typography>
-
-                            <Typography
-                                variant={isXs ? "body1" : "h6"}
-                                mb={3}
-                                sx={{
-                                    color: "rgba(255,255,255,0.9)",
-                                    textShadow: "0 0 6px rgba(0,0,0,.45)",
-                                }}
-                            >
-                                Начните собирать свою команду уже сегодня!
-                            </Typography>
-
-                            <AuthButtons loggedIn={isLoggedIn} />
-                        </Box>
-                    </Container>
-                </Box>
-            </Box>
-            <Box
-                sx={{
-                    position: "absolute",
-                    bottom: 20,
-                    left: "50%",
-                    transform: "translateX(-50%)",
-                    zIndex: 2,
-                    color: "white",
-                    filter: "drop-shadow(0 0 12px rgba(0,0,0,0.8))",
-                    opacity: 0.9,
-                    animation: "bounce 2s infinite",
-                }}
-            >
-                <KeyboardArrowDownIcon sx={{ fontSize: 80 }} />
-            </Box>
-            <style>
-                {`
-      @keyframes bounce {
-        0%, 20%, 50%, 80%, 100% { transform: translateX(-50%) translateY(0); }
-        40% { transform: translateX(-50%) translateY(12px); }
-        60% { transform: translateX(-50%) translateY(6px); }
-      }
-    `}
-            </style>
-
-            {/* События */}
-            <Box
-                sx={{
-                    py: 7,
-                    bgcolor: "grey.900",
-                    color: "white",
-                }}
-            >
-                <Container maxWidth="lg">
-                    <Typography
-                        variant="h4"
-                        align="center"
-                        gutterBottom
-                        sx={{
-                            fontWeight: 700,
-                            // textShadow: "0 0 8px rgba(254, 221, 44,.35)",
-                            mb: 4,
-                        }}
-                    >
-                        Последние события
-                    </Typography>
-
-                    {loading ? (
-                        <Stack spacing={3}>
-                            {Array.from({ length: 3 }).map((_, idx) => (
-                                <Box key={idx}>{renderSkeletonCard()}</Box>
-                            ))}
-                        </Stack>
-                    ) : events.length === 0 ? (
-                        <Typography>Нет доступных событий</Typography>
-                    ) : (
-                        <>
-                            <Grid container spacing={3} justifyContent="center"
-                                  sx={{ display: { xs: 'none', sm: 'flex' } }}
-                            >
-                                {events.slice(0, 4).map((event) => {
-                                const participation = participationMap[event.id];
-                                const isParticipating = !!participation;
-                                    return (
-                                        <Grid item
-                                          xs={12}
-                                          sm={6}
-                                          md={4}
-                                          lg={3}
-                                          key={event.id}
-                                          sx={{ display: 'flex' }}>
-                                        <Card
-                                            onClick={() =>
-                                                navigate(`/event/${event.id}`, {
-                                                    state: participation ? { participant_id: participation.id } : undefined
-                                                })
-                                            }
-                                            sx={{
-                                                width: "100%",
-                                                height: "100%",
-                                                display: "flex",
-                                                flexDirection: "column",
-                                                bgcolor: "background.paper",
-                                                color: "text.primary",
-                                                borderRadius: 2,
-                                                transition: "transform .2s, box-shadow .2s",
-                                                "&:hover": {
-                                                    transform: "translateY(-4px)",
-                                                    boxShadow: 6,
-                                                },
-                                                cursor: "pointer",
-                                                flexGrow: 1,
-                                            }}
-                                        >
-                                            <CardContent sx={{width:'100%'}}>
-                                                <Typography variant="h6" gutterBottom fontWeight={700}>
-                                                    {event.name}
-                                                </Typography>
-                                                <Typography
-                                                    variant="body2"
-                                                    color="text.secondary"
-                                                    sx={{
-                                                        display: "-webkit-box",
-                                                        WebkitLineClamp: 3,
-                                                        WebkitBoxOrient: "vertical",
-                                                        overflow: "hidden",
-                                                        mb: 1,
-                                                    }}
-                                                >
-                                                    {event.description}
-                                                </Typography>
-                                                <Typography variant="body2" sx={{ mt: 0.5 }}>
-                                                    <b>Дата начала:</b> {event.start_date}
-                                                </Typography>
-                                                <Typography variant="body2">
-                                                    <b>Дата окончания:</b> {event.end_date}
-                                                </Typography>
-
-                                                <Box mt={1}>
-                                                    <Typography variant="subtitle2">Треки:</Typography>
-                                                    <Stack direction="row" spacing={1} flexWrap="wrap" rowGap={1}>
-                                                        {(event.event_tracks || []).map((track) => (
-                                                            <Chip
-                                                                key={track.id}
-                                                                label={track.name}
-                                                                size="small"
-                                                                variant="outlined"
-                                                                sx={{ mb: 0.5 }}
-                                                            />
-                                                        ))}
-                                                    </Stack>
-                                                </Box>
-                                            </CardContent>
-                                            <Box sx={{ mt: "auto", p: 2 }}>
-                                            {isParticipating ? (
-                                                <Typography
-                                                    sx={{
-                                                        mt: 3,
-                                                        p: 1,
-                                                        bgcolor: grey[800],
-                                                        borderRadius: 1,
-                                                        textAlign: "center",
-                                                        color: "white",
-                                                        width: "100%"
-                                                    }}
-                                                >
-                                                    Вы уже участвуете <br />
-                                                    {participation.event_role === "TEAMLEAD"
-                                                        ? "Тимлид"
-                                                        : "Участник"}
-                                                    {" • "}
-                                                    {participation.track?.name}
-                                                </Typography>
-                                            ) : (
-                                                <Button
-                                                    variant="contained"
-                                                    fullWidth
-                                                    onClick={(e) => {
-                                                        e.stopPropagation();
-                                                        navigate(`/event/${event.id}/register`);
-                                                    }}
-                                                >
-                                                    Я участвую
-                                                </Button>)}
-                                                </Box>
-                                        </Card>
-                                    </Grid>
-                                )})}
-                            </Grid>
-                            <Box mt={4} display="flex" justifyContent="center">
-                                <Button
-                                    variant="contained"
-                                    color="primary"
-                                    onClick={() => navigate("/events")}
-                                    endIcon={<ArrowForwardIosIcon />}
-                                >
-                                    Смотреть все события
-                                </Button>
-                            </Box>
-                        </>
-                    )}
-                </Container>
-            </Box>
-
-            {/* Какие мероприятия мы поддерживаем */}
-            <Box
-                sx={{
-                    py: 7,
-                    bgcolor: "background.papaer",
-                }}
-            >
-                <Box
-                    sx={{
-                        display: "flex",
-                        justifyContent: "center",
-                        mb: 2
-                    }}
-                >
-                    <Box
-                        component="img"
-                        src="logo_yellow.png"
-                        sx={{
-                            width: { xs: "35%", sm: "240px" }, // меньше на мобилке
-                            height: "auto"
-                        }}
-                    />
-                </Box>
-                <Container maxWidth="lg">
-                    <Typography
-                        variant="h4"
-                        align="center"
-                        gutterBottom
-                        sx={{
-                            fontWeight: 700,
-                            mb: 4,
-                            fontSize: {
-                                xs: "1.5rem",
-                                sm: "2rem",
-                                md: "2.5rem",
-                            },
-                        }}
-                    >
-                        Единый сервис для всех олимпиад и хакатонов
-                    </Typography>
-
-                    <Grid container spacing={3} justifyContent="center">
-                        {[
-                            {
-                                logo: "prod.png",
-                                title: "",
-                                color: '#003f28',
-                                url: "https://prodcontest.ru/"
-                            },
-                            {
-                                logo: "dano.png",
-                                title: "",
-                                color: '#323332',
-                                url: "https://dano.hse.ru/"
-                            },
-                            {
-                                logo: "deadline.png",
-                                title: "",
-                                color: '#141414',
-                                url: "https://event.centraluniversity.ru/casecontest"
-                            },
-                        ].map((event, idx) => (
-                            <Grid item xs={12} sm={6} md={4} key={idx}>
-                                <Stack
-                                    alignItems="center"
-                                    justifyContent="center"
-                                    spacing={2}
-                                    sx={{
-                                        p: 3,
-                                        borderRadius: 2,
-                                        bgcolor: event.color,
-                                        border: (t) =>
-                                            `1px solid ${
-                                                t.palette.mode === "dark"
-                                                    ? "rgba(255,255,255,0.08)"
-                                                    : "rgba(0,0,0,0.08)"} `,
-                                        height: { xs: "140px", sm: "100%" },
-                                        textAlign: "center",
-                                    }}
-                                >
-                                    {/* Логотип */}
-                                    <Box
-                                        component="a"
-                                        href={event.url}
-                                        target="_blank"
-                                        rel="noopener noreferrer"
-                                        sx={{ display: "block", width: "100%" }}
-                                    >
-                                        <Box
-                                            component="img"
-                                            src={event.logo}
-                                            //alt={event.title}
-                                            sx={{
-                                                width: {xs: "100%", sm: 300},
-                                                height: {sm: "auto", xs: "auto"}
-                                            }}
-                                        />
-                                    </Box>
-                                </Stack>
-                            </Grid>
-                        ))}
-                    </Grid>
-                </Container>
-            </Box>
-
-            {/* FAQ / Помощь пользователю */}
-            <Box sx={{ py: 7, bgcolor: "grey.900", color: "text.primary" }}>
-                <Container maxWidth="md">
-                    <Typography
-                        variant="h4"
-                        align="center"
-                        gutterBottom
-                        sx={{ fontWeight: 700, mb: 4 }}
-                    >
-                        Часто задаваемые вопросы (FAQ)
-                    </Typography>
-
-                    <Stack>
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как начать пользоваться FindMyTeam?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Перейдите на главную страницу, создайте аккаунт или войдите. После входа вы сможете выбрать мероприятие, зарегистрироваться и указать свою роль — участник или тимлид.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как зарегистрироваться на мероприятие?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    В разделе “Мероприятия” выберите интересующее событие и нажмите “Я участвую”.
-                                    Укажите резюме, выберите трек и роль. Если вы тимлид, создайте команду и добавьте вакансии. Вы также можете добавить или убрать вакансии позже на странице команды. В резюме, описании команд и описании вакансий поддерживается Markdown разметка.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как изменить резюме, трек или роль на мероприятии?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Для этого перейдите на страницу мероприятия, затем нажмите на иконку "Карандаш" рядом с данными, которые хотите редактировать (данные об участии находятся внизу страницы). Вы можете изменить резюме и трек, а также отказаться от участия в мероприятии.
-                                    Чтобы сменить роль с "Тимлид" на "Участник", выйдете из команды или удалите команду. Если тимлид выходит из команды, команда автоматически удаляется, а тимлид становится обычным участником.
-                                    Чтобы сменить роль с "Участник" на "Тимлид" откажитесь от участия в мероприятии, а затем зарегистрируйтесь заново, выбрав роль "Тимлид" на этапе регистрации. Обратите внимание, что в таком случае все данные об участии в мероприятии (резюме и трек) удалятся.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как собрать команду?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Тимлид создаёт команду при регистрации, добавляет вакансии и может приглашать участников через вкладку “Участники”.
-                                    Участники могут откликаться на вакансии команд с подходящим треком.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как посмотреть свои мероприятия?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Перейдите в личный кабинет. Там отображаются все мероприятия, где вы участвуете, с вашими ролями и треками.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Что делать, если я не вижу подходящую команду?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Можно создать собственную команду, если вы тимлид, или подождать, пока откроются новые вакансии.
-                                    Также вы можете откликнуться на несколько подходящих вакансий сразу.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как отправить приглашение?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Отправить приглашение можно только если вы тимлид и в вашей команде есть открытые вакансии.
-                                    Откройте список участников на странице мероприятия, нажмите на кнопку "пригласить" на карточке участника, а затем выберите вакансию, на которую вы хотите его пригласить.
-                                    Посмотреть приглашения, которые вы отправили можно во вкладке "Мои приглашения" на странице мероприятия.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как принять приглашение в команду?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Все приглашения отображаются во вкладках “Приглашения мне” на странице мероприятия.
-                                    Вы можете принять или отклонить приглашения в один клик.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как посмотреть участников команды?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Чтобы посмотреть подробную информацию о команде, откройте вкладку "команды" на странице мероприятий. Далее нажмите на строку с командой, чтобы перейти на страницу этой команды.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как управлять своей командой?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Управление командной доступно только если вы тимлид в этой команде. Для этого перейдите на страницу с вашей командой. Там вы можете редактировать название и описание команды, вы также можете добавлять и удалять вакансии.
-                                    Для удаления вакансий нажмите на иконку мусорной корзины на карточке вакансии. Для удаления участников из команды нажмите на кнопку "выгнать" на карточке участника.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как откликнуться на вакансию?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Откликнуться на вакансию можно только если вы участник мероприятия и у вас еще нет команды.
-                                    Чтобы откликнуться на вакансию, откройте вкладку "вакансии" на странице мероприятия. Далее нажмите "откликнуться" на карточке вакансии, которая вам интересна.
-                                    Вы также можете это сделать со страницы команды.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как редактировать данные в профиле или выйти из аккаунта?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    В личном кабинете вы можете изменить имя, e-mail, пароль и Telegram.
-                                    Там же можно выйти или удалить профиль. Если вы хотите изменить резюме или отказаться от участия в каком-то мероприятии, вы можете сделать это на странице мероприятия.
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-                        <Accordion>
-                            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                                <Typography variant="subtitle1" fontWeight={600}>
-                                    Как создавать мероприятия и управлять ими (для организаторов)?
-                                </Typography>
-                            </AccordionSummary>
-                            <AccordionDetails>
-                                <Typography variant="body2">
-                                    Панель администратора предназначена только для организаторов мероприятий.
-                                    Если вы организатор, нажмите {' '}
-                                    <Link
-                                        href="https://findmyteam.ru/admin"
-                                        target="_blank"
-                                        rel="noopener noreferrer"
-                                        sx={{ textDecoration: 'none' }}
-                                    >
-                                        здесь
-                                    </Link>
-                                     . Далее введите логин и пароль от учетной записи администратора. Затем в боковой панели выберите категорию, которую хотите изменить или посмотреть (Мероприятия, треки, команды и тд).
-                                </Typography>
-                            </AccordionDetails>
-                        </Accordion>
-
-                    </Stack>
-                </Container>
-            </Box>
-
-
-            {/* Footer */}
-            <Box sx={{ bgcolor: "grey.800", color: "white", py: 4 }}>
-                <Container>
-                    <Stack
-                        direction={{ xs: "column", sm: "row" }}
-                        justifyContent="space-between"
-                        spacing={2}
-                        alignItems={{ xs: "flex-start", sm: "center" }}
-                    >
-                        <Typography variant="body2">© 2025 FindMyTeam</Typography>
-                    </Stack>
-                </Container>
-            </Box>
+            <HeroSection isLoggedIn={isLoggedIn} />
+            <EventsSection 
+                events={events} 
+                loading={loading} 
+                participationMap={participationMap}
+            />
+            <SupportedEventsSection />
+            <FAQSection />
+            <Footer />
         </>
     );
 }
